Basic_Lab/requirements.txt
fastapi
uvicorn
motor
python-jose
passlib[bcrypt]
pydantic
python-dotenv
